requirements.txt
# Core ML and NLP dependencies
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
accelerate>=0.20.0

# OCR dependencies
paddlepaddle>=2.5.0
paddleocr>=2.7.0
opencv-python>=4.8.0
Pillow>=9.5.0

# API and HTTP requests
requests>=2.31.0
httpx>=0.24.0

# Data processing and validation
numpy>=1.24.0
pandas>=2.0.0
pydantic>=2.0.0
python-dateutil>=2.8.0

# Evaluation and metrics
scikit-learn>=1.3.0
seaborn>=0.12.0
matplotlib>=3.7.0

# Development and testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0

# Documentation
sphinx>=7.0.0
sphinx-rtd-theme>=1.3.0

# Optional: For advanced OCR (requires NVIDIA API key)
# nvidia-ml-py3>=7.352.0
